+++ engine/src/test/java/net/jqwik/engine/properties/shrinking/ShrinkableTypesForTest.java
@@ -73,7 +73,7 @@ public class ShrinkableTypesForTest {
 		@Override
 		public Stream<Shrinkable<Integer>> shrink() {
 			try {
-				Thread.sleep(500);
+				Thread.sleep(100);
 			} catch (InterruptedException e) {
 				e.printStackTrace();
 			}