+++ docs/release-notes.md
@@ -28,6 +28,8 @@ title: jqwik Release Notes
 
 - Added `Tuple.of()` and `Tuple.empty()` to create an empty tuple.
 
+- The time out for bounded shrinking can now be changed in `jqwik.properties`
+
 #### Breaking Changes
 
 - Shrinking is no longer bound by number of shrinking attempts, but by time with a 10 seconds default.